requirements.txt
Django==2.1.3
django-filter==2.0.0
djangorestframework==3.9.0
gunicorn==19.9.0
Markdown==3.0.1
psycopg2==2.7.6.1
django-cors-headers
